move `getWaypoint` to RouteUtils
